Update README.md
In this CA we were assigned to create a web service to convert numbers to Roman numerals and vice versa. The second part of this CA was to create an API proxy to interact with the Open weather API. 

Technologies : Java, Maven, Jersey, NetBeans IDE, "javax.ws.rs" library, and "org.apache.http" library.